# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RT Finance Backend API",
        "description": "API Documentation for RT Finance Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Local Development"
        },
        {
            "url": "https://rt-finance-backend.vercel.app/api",
            "description": "Production"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/monthly-fee/breakdown/{year}/{month}": {
            "get": {
                "tags": [
                    "Monthly Fee"
                ],
                "summary": "Get monthly breakdown report",
                "parameters": [
                    {
                        "name": "year",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "month",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Breakdown data"
                    }
                }
            }
        },
        "/monthly-fee/pending-submission": {
            "get": {
                "tags": [
                    "Monthly Fee - RW Submission"
                ],
                "summary": "List fees pending RW submission",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grouped pending fees"
                    }
                }
            }
        },
        "/monthly-fee/submit-to-rw": {
            "post": {
                "tags": [
                    "Monthly Fee - RW Submission"
                ],
                "summary": "Submit fees to RW",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "period": {
                                        "type": "string",
                                        "example": "2026-01"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "ids",
                                    "period"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Submission created"
                    }
                }
            }
        },
        "/monthly-fee/rw-submissions": {
            "get": {
                "tags": [
                    "Monthly Fee - RW Submission"
                ],
                "summary": "List all RW submissions",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of submissions"
                    }
                }
            }
        },
        "/monthly-fee/rw-submissions/{id}": {
            "get": {
                "tags": [
                    "Monthly Fee - RW Submission"
                ],
                "summary": "Get RW submission detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission detail with isLate flag"
                    }
                }
            }
        },
        "/monthly-fee/history": {
            "get": {
                "tags": [
                    "Monthly Fee"
                ],
                "summary": "Payment history for resident(s)",
                "parameters": [
                    {
                        "name": "startYear",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startMonth",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "endYear",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "endMonth",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "block",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "houseNumber",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment history per month"
                    }
                }
            }
        },
...
    }
}